fix: [container]  Avoid 404 errors on apt install
With Docker layer caching, it can pass a long time between to two
layers invocation. This means that `apt-get install` can receive 404
errors when run on a different layer than `apt-get update`. So, we can
either run `apt-get update` before the upgrade and before the install,
or, to avoid duplicating invocation, run the three commands on the same
layer.